+++ modules/admin/views/news-category/index.php
@@ -1,5 +1,4 @@
 <?php
-
 /**
  * @var $this yii\web\View
  * @var $widget kartik\grid\GridView
@@ -7,21 +6,22 @@
 
 use yii\helpers\Html;
 
-$this->title = 'Users';
+$this->title = 'News Categories';
 $this->params['breadcrumbs'] = [
     ['label' => 'Admin Panel', 'url' => ['statistic/index']],
     $this->title,
 ];
 ?>
-<div class="admin-user">
+<div class="admin-news-category">
     <div class="row">
-        <div class="col-sm-10">
+        <div class="col-sm-9">
             <h3><?= Html::encode($this->title); ?></h3>
         </div>
-        <div class="col-sm-2">
-            <?= Html::a('Create', ['user/create'], ['class' => 'float-right btn btn-success']); ?>
+        <div class="col-sm-3">
+            <?= Html::a('Back to News', ['news/index'], ['class' => 'btn btn-info']); ?>
+            <?= Html::a('Create', ['news-category/create'], ['class' => 'float-right btn btn-success']); ?>
         </div>
     </div>
-    <?php // echo $this->render('_search', ['model' => $searchModel]); ?>
+    <p></p>
     <?= $widget->run(); ?>
 </div>